+++ src/applications/mhv-secure-messaging/tests/components/Compose/ComposeForm.unit.spec.jsx
@@ -973,4 +973,217 @@ describe('Compose form component', () => {
       expect(checkbox).to.have.attribute('error', '');
     });
   });
+
+  it('should display an error message when a file is 0B', async () => {
+    const screen = setup(initialState, Paths.COMPOSE);
+    const file = new File([''], 'test.png', { type: 'image/png' });
+
+    const uploader = screen.getByTestId('attach-file-input');
+
+    await waitFor(() =>
+      fireEvent.change(uploader, {
+        target: { files: [file] },
+      }),
+    );
+
+    const error = screen.getByTestId('file-input-error-message');
+    expect(error).to.exist;
+    expect(error.textContent).to.equal(
+      ErrorMessages.ComposeForm.ATTACHMENTS.FILE_EMPTY,
+    );
+  });
+
+  it('should display an error message when a file is not an accepted file type', async () => {
+    const file = new File(['(⌐□_□)'], 'test.zip', {
+      type: 'application/zip',
+    });
+    const screen = setup(initialState, Paths.COMPOSE);
+
+    const uploader = screen.getByTestId('attach-file-input');
+
+    await waitFor(() =>
+      fireEvent.change(uploader, {
+        target: { files: [file] },
+      }),
+    );
+
+    const error = screen.getByTestId('file-input-error-message');
+    expect(error).to.exist;
+    expect(error.textContent).to.equal(
+      ErrorMessages.ComposeForm.ATTACHMENTS.INVALID_FILE_TYPE,
+    );
+  });
+
+  it('should display an error message when a file is a duplicate', async () => {
+    const file = new File(['(⌐□_□)'], 'test.png', { type: 'image/png' });
+    const customDraftMessage = {
+      ...draftMessage,
+      recipientId: 1013155,
+      recipientName: '***MEDICATION_AWARENESS_100% @ MOH_DAYT29',
+      triageGroupName: '***MEDICATION_AWARENESS_100% @ MOH_DAYT29',
+      attachments: [file],
+    };
+
+    const customState = {
+      ...draftState,
+      sm: {
+        ...draftState.sm,
+        draftDetails: { customDraftMessage },
+      },
+    };
+
+    const screen = renderWithStoreAndRouter(
+      <ComposeForm
+        draft={customDraftMessage}
+        recipients={customState.sm.recipients}
+      />,
+      {
+        initialState: customState,
+        reducers: reducer,
+        path: `/draft/${draftMessage.id}`,
+      },
+    );
+    const uploader = screen.getByTestId('attach-file-input');
+    await waitFor(() =>
+      fireEvent.change(uploader, {
+        target: { files: [file] },
+      }),
+    );
+    const error = screen.getByTestId('file-input-error-message');
+    expect(error.textContent).to.equal('You have already attached this file.');
+  });
+
+  it('should display an error message when a file with the same name but different size is a duplicate', async () => {
+    const oneAttachment = { name: 'test.png', size: 100, type: 'image/png' };
+
+    const file = new File(['(⌐□_□)'], 'test.png', {
+      type: 'image/png',
+      size: 200,
+    });
+
+    const customDraftMessage = {
+      ...draftMessage,
+      recipientId: 1013155,
+      recipientName: '***MEDICATION_AWARENESS_100% @ MOH_DAYT29',
+      triageGroupName: '***MEDICATION_AWARENESS_100% @ MOH_DAYT29',
+      attachments: [oneAttachment],
+    };
+
+    const customState = {
+      ...draftState,
+      sm: {
+        ...draftState.sm,
+        draftDetails: { customDraftMessage },
+      },
+    };
+
+    const screen = renderWithStoreAndRouter(
+      <ComposeForm
+        draft={customDraftMessage}
+        recipients={customState.sm.recipients}
+      />,
+      {
+        initialState: customState,
+        reducers: reducer,
+        path: `/draft/${draftMessage.id}`,
+      },
+    );
+
+    const uploader = screen.getByTestId('attach-file-input');
+
+    await waitFor(() =>
+      fireEvent.change(uploader, {
+        target: { files: [file] },
+      }),
+    );
+
+    expect(uploader.files[0].name).to.equal('test.png');
+    expect(uploader.files.length).to.equal(1);
+    const error = screen.getByTestId('file-input-error-message');
+    expect(error.textContent).to.equal('You have already attached this file.');
+  });
+
+  it('should display an error message when a file is over 6MB', async () => {
+    const largeFileSizeInBytes = 7 * 1024 * 1024; // 7MB
+
+    const largeFileBuffer = new ArrayBuffer(largeFileSizeInBytes);
+    const largeFileBlob = new Blob([largeFileBuffer], {
+      type: 'application/octet-stream',
+    });
+
+    const largeFile = new File([largeFileBlob], 'large_file.txt', {
+      type: 'application/octet-stream',
+      lastModified: new Date().getTime(),
+    });
+
+    const screen = setup(initialState, Paths.COMPOSE);
+    const uploader = screen.getByTestId('attach-file-input');
+    await waitFor(() =>
+      fireEvent.change(uploader, {
+        target: { files: [largeFile] },
+      }),
+    );
+    const error = screen.getByTestId('file-input-error-message');
+    expect(error.textContent).to.equal(
+      ErrorMessages.ComposeForm.ATTACHMENTS.FILE_TOO_LARGE,
+    );
+  });
+
+  it(' should display an error message when attaching a new file increases total attachments size over 10MB', async () => {
+    const oneMB = 1024 * 1024;
+    const customAttachments = [
+      { name: 'test1.png', size: 4 * oneMB, type: 'image/png' },
+      { name: 'test2.png', size: 4 * oneMB, type: 'image/png' },
+    ];
+    const largeFileSizeInBytes = 3 * oneMB; // 7MB
+
+    const largeFileBuffer = new ArrayBuffer(largeFileSizeInBytes);
+    const largeFileBlob = new Blob([largeFileBuffer], {
+      type: 'application/octet-stream',
+    });
+
+    const largeFile = new File([largeFileBlob], 'large_file.txt', {
+      type: 'application/octet-stream',
+      lastModified: new Date().getTime(),
+    });
+
+    const customDraftMessage = {
+      ...draftMessage,
+      recipientId: 1013155,
+      recipientName: '***MEDICATION_AWARENESS_100% @ MOH_DAYT29',
+      triageGroupName: '***MEDICATION_AWARENESS_100% @ MOH_DAYT29',
+      attachments: [...customAttachments],
+    };
+
+    const customState = {
+      ...draftState,
+      sm: {
+        ...draftState.sm,
+        draftDetails: { customDraftMessage },
+      },
+    };
+
+    const screen = renderWithStoreAndRouter(
+      <ComposeForm
+        draft={customDraftMessage}
+        recipients={customState.sm.recipients}
+      />,
+      {
+        initialState: customState,
+        reducers: reducer,
+        path: `/draft/${draftMessage.id}`,
+      },
+    );
+
+    const uploader = screen.getByTestId('attach-file-input');
+    await waitFor(() =>
+      fireEvent.change(uploader, {
+        target: { files: [largeFile] },
+      }),
+    );
+    const error = screen.getByTestId('file-input-error-message');
+    expect(error.textContent).to.equal(
+      ErrorMessages.ComposeForm.ATTACHMENTS.TOTAL_MAX_FILE_SIZE_EXCEEDED,
+    );
+  });
 });
